win32.rapi: Use native D interfaces instead of Vtbl structs
Since D interfaces are reference types, this module was an exception in that it defined an IRapiStream type which was a value type. This required using "IRapiStream*" in places where normally one would write "IRapiStream".

This is a breaking change, although I doubt anyone actually uses this module (it's related to Windows CE).

git-svn-id: http://svn.dsource.org/projects/bindings/trunk@407 e3080b08-b6e0-0310-b0a6-cbd5e3aa05fd