Rename uuid.di to uuid.d
The win32.uuid module contains value-constants (not manifest constants), and as such needs to be compiled and linked. Thus, the .di extension is in appropriate - it causes the module to not be included in the library (as built by the makefile), and also causes rdmd to ignore the file (not include it in dmd's command line), which results in linker errors.

git-svn-id: http://svn.dsource.org/projects/bindings/trunk@410 e3080b08-b6e0-0310-b0a6-cbd5e3aa05fd